move to keras directory
