+++ efficientdet/efficientdet_arch_test.py
@@ -21,7 +21,6 @@ from __future__ import print_function
 import tensorflow.compat.v1 as tf
 
 import efficientdet_arch
-import efficientdet_arch_keras
 import hparams_config
 import utils
 
@@ -93,31 +92,6 @@ class EfficientDetArchTest(tf.test.TestCase):
     self.assertSequenceEqual((51871782, 324815496167),
                              self.build_model('efficientdet-d7', 1536))
 
-class KerasTest(tf.test.TestCase):
-  def test_resample_feature_map(self):
-    feat = tf.random.uniform([1, 16, 16, 320])
-    for apply_fn in [True, False]:
-      for is_training in [True, False]:
-        for use_tpu in [True, False]:
-          with self.subTest(apply_fn=apply_fn,is_training=is_training, use_tpu=use_tpu):
-            tf.random.set_random_seed(111111)
-            expect_result = efficientdet_arch.resample_feature_map(feat,
-                                                                   name='resample_p0',
-                                                                   target_height=8,
-                                                                   target_width=8,
-                                                                   target_num_channels=64,
-                                                                   apply_bn=apply_fn,
-                                                                   is_training=is_training,
-                                                                   use_tpu=use_tpu)
-            tf.random.set_random_seed(111111)
-            actual_result = efficientdet_arch_keras.ResampleFeatureMap(name='resample_p0',
-                                                                       target_height=8,
-                                                                       target_width=8,
-                                                                       target_num_channels=64,
-                                                                       apply_bn=apply_fn,
-                                                                       is_training=is_training,
-                                                                       use_tpu=use_tpu)(feat)
-            self.assertAllCloseAccordingToType(expect_result, actual_result)
 
 class EfficientDetArchPrecisionTest(tf.test.TestCase):
